# specs/001-agent-service-cpq/contracts/openapi.template.yaml
openapi: 3.1.0
info:
  title: ClientWeave Public and Agent API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Generated external contract for the ordinary buyer UI and six page-provided
    agent capabilities. Authoritative executable schemas live under
    src/contracts/schemas/. Owner administration uses the same domain services
    behind authenticated server actions and is described by the plan and data model.
servers:
  - url: /api/v1
...
paths:
...
  /scopes/current:
    get:
      tags: [Scope]
      operationId: get_current_scope
      summary: Read the one scope authorized by an agent bearer or browser session
      security:
        - scopeBearer: []
        - scopeSessionCookie: []
      responses:
        '200':
          description: Current review model
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScopeReview' }
        '404': { $ref: '#/components/responses/ScopeUnavailable' }
    patch:
      tags: [Scope]
      operationId: update_scope
      summary: Apply validated changes to the shared draft
      security:
        - scopeBearer: []
        - scopeSessionCookie: []
          csrfHeader: []
      x-agent-capability:
        stateEffect: DRAFT_MUTATION
        consequential: false
      parameters:
        - in: header
          name: If-Match
          required: true
          description: >-
            Canonical HTTP concurrency precondition containing the quoted last-seen
            scope revision, for example `"7"`. The WebMCP adapter translates its
            expected_revision argument into this header.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateScopeRequest' }
      responses:
        '200':
          description: Updated scope and validation state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScopeReview' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/ScopeUnavailable' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /scopes/current/human-confirmations:
    post:
      tags: [Confirmation]
      operationId: record_human_confirmation
      summary: Record a direct human confirmation of the current final summary
      description: This operation is intentionally not registered as an agent capability and rejects bearer authentication.
      security:
        - scopeSessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [summaryNonce]
              properties:
                summaryNonce: { type: string, minLength: 32, maxLength: 512 }
              additionalProperties: false
      responses:
        '201':
          description: Current human confirmation
          content:
            application/json:
              schema:
                type: object
                required: [confirmationId, confirmedAt]
                properties:
                  confirmationId: { type: string, format: uuid }
                  confirmedAt: { type: string, format: date-time }
                additionalProperties: false
        '403': { $ref: '#/components/responses/HumanGestureRequired' }
        '409': { $ref: '#/components/responses/Conflict' }
  /scopes/current/finalizations:
    post:
      tags: [Finalization]
      operationId: finalize_confirmed_scope
      summary: Create the single lead and optional booking from a current confirmation
      security:
        - scopeBearer: []
        - scopeSessionCookie: []
          csrfHeader: []
      x-agent-capability:
        stateEffect: CONSEQUENTIAL_WRITE
        consequential: true
        requiresCurrentHumanConfirmation: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmationId]
              properties:
                confirmationId: { type: string, format: uuid }
              additionalProperties: false
      responses:
        '200':
          description: Existing result returned for a successful retry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FinalizationResult' }
        '201':
          description: Lead and optional booking created atomically
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FinalizationResult' }
        '404': { $ref: '#/components/responses/ScopeUnavailable' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    scopeBearer:
      type: http
      scheme: bearer
      bearerFormat: OpaqueScopeToken
      description: Agent credential authorizing only one unexpired scope; stored hashed server-side.
    scopeSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-clientweave_scope
      description: Secure HttpOnly browser session authorizing only one unexpired scope.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the scope session cookie for browser mutations.
  parameters:
    IdempotencyKeyRequired:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
  responses:
...
    ValidationError:
      description: Input or current-state validation failed safely
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Conflict:
      description: Revision, quote, confirmation, idempotency, or slot conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    ScopeUnavailable:
      description: Scope credential is invalid, expired, revoked, or does not authorize this resource
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    PreconditionFailed:
      description: The supplied scope revision is stale
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    PreconditionRequired:
      description: A required revision precondition is missing
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    HumanGestureRequired:
      description: Confirmation did not originate from the trusted human UI boundary
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    RateLimited:
      description: Public request limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    MoneyRange:
      type: object
      required: [currency, minMinor, maxMinor]
      properties:
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        minMinor: { type: integer, minimum: 0 }
        maxMinor: { type: integer, minimum: 0 }
      additionalProperties: false
    ServiceMatch:
      type: object
      required: [serviceId, slug, name, description, basePrice, deliveryBusinessDays, includedItems, addOns, intakeFields, constraints, fit]
      properties:
        serviceId: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        basePrice: { $ref: '#/components/schemas/MoneyRange' }
        deliveryBusinessDays:
          type: object
          required: [min, max]
          properties:
            min: { type: integer, minimum: 1 }
            max: { type: integer, minimum: 1 }
          additionalProperties: false
        includedItems: { type: array, items: { type: string } }
        addOns: { type: array, items: { type: string } }
        intakeFields:
          type: array
          items: { $ref: '#/components/schemas/ScopeField' }
        constraints:
          type: array
          items: { $ref: '#/components/schemas/ServiceConstraintSummary' }
        fit:
          type: object
          required: [eligible, reasons, conflicts]
          properties:
            eligible: { type: boolean }
            reasons: { type: array, items: { type: string } }
            conflicts: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
          additionalProperties: false
      additionalProperties: false
    ScopeField:
      type: object
      required: [key, label, valueType, required, priceAffecting]
      properties:
        key: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        label: { type: string }
        helpText: { type: string }
        valueType: { enum: [TEXT, INTEGER, DECIMAL, BOOLEAN, DATE, SINGLE_CHOICE, MULTI_CHOICE] }
        required: { type: boolean }
        priceAffecting: { type: boolean }
        allowedValues: { type: array, items: { type: [string, number, boolean] } }
      additionalProperties: false
...
    UpdateScopeRequest:
      type: object
      required: [changes]
      properties:
        changes:
          type: object
          properties:
            goal: { type: string, minLength: 1, maxLength: 2000 }
            budgetMaxMinor: { type: [integer, 'null'], minimum: 0 }
            targetDeliveryDate: { type: [string, 'null'], format: date }
            assumptions:
              type: array
              maxItems: 20
              uniqueItems: true
              items: { type: string, minLength: 1, maxLength: 500 }
            answers:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/AnswerInput' }
          additionalProperties: false
      additionalProperties: false
    AnswerInput:
      type: object
      required: [fieldKey, value]
      properties:
        fieldKey: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        value: {}
      additionalProperties: false
    ScopeReview:
      type: object
      required: [scopeRef, status, revision, service, goal, budgetMaxMinor, targetDeliveryDate, assumptions, answers, missingFields, conflicts, quoteStatus, expiresAt]
      properties:
        scopeRef: { type: string, minLength: 16, maxLength: 128 }
        status: { enum: [DRAFT, FINALIZED, EXPIRED] }
        revision: { type: integer, minimum: 1 }
        service: { $ref: '#/components/schemas/ServiceMatch' }
        goal: { $ref: '#/components/schemas/AttributedTextValue' }
        budgetMaxMinor: { $ref: '#/components/schemas/AttributedNullableMoneyValue' }
        targetDeliveryDate: { $ref: '#/components/schemas/AttributedNullableDateValue' }
        assumptions:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/AttributedTextValue' }
        answers:
          type: array
          items: { $ref: '#/components/schemas/AttributedAnswer' }
        missingFields: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
        conflicts: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
        quoteStatus: { enum: [NONE, CURRENT, STALE, INCOMPLETE, CONFLICTED] }
        expiresAt: { type: string, format: date-time }
        continuationUrl:
          type: string
          format: uri
          description: Returned only once from scope creation; its secret is in the URL fragment.
      additionalProperties: false
    ValidationIssue:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        fieldKey: { type: string }
        message: { type: string }
        severity: { enum: [INFO, WARNING, ERROR], default: ERROR }
        suggestedActions: { type: array, items: { type: string } }
      additionalProperties: false
    ServiceConstraintSummary:
      type: object
      required: [type, fieldKeys, message]
      properties:
        type: { enum: [DELIVERY_MINIMUM, INCOMPATIBLE_OPTIONS, FIELD_LIMIT, BUDGET_FLOOR] }
        fieldKeys:
          type: array
          items: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        message: { type: string, minLength: 1, maxLength: 500 }
      additionalProperties: false
    ProvenanceMetadata:
      type: object
      required: [actorType, updatedAt]
      properties:
        actorType: { enum: [HUMAN, AGENT, IMPORTED, SYSTEM] }
        actorId: { type: [string, 'null'], maxLength: 255 }
        updatedAt: { type: string, format: date-time }
      additionalProperties: false
    AttributedTextValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: string }
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
    AttributedNullableMoneyValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: [integer, 'null'], minimum: 0 }
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
    AttributedNullableDateValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: [string, 'null'], format: date }
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
    AttributedAnswer:
      type: object
      required: [fieldKey, value, provenance]
      properties:
        fieldKey: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        value: {}
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
...
    FinalizationResult:
      type: object
      required: [scopeRef, leadId, status, finalizedAt]
      properties:
        scopeRef: { type: string, minLength: 16, maxLength: 128 }
        leadId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        status: { enum: [FINALIZED] }
        finalizedAt: { type: string, format: date-time }
      additionalProperties: false
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, code, retryable, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        retryable: { type: boolean }
        fieldErrors: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
        currentRevision: { type: integer, minimum: 1 }
        traceId: { type: string }